requirements.txt
pymongo==4.1.1
Flask-PyMongo==2.3.0
Flask==2.1.2
python-dotenv==0.20.0
requests==2.28.0
Werkzeug~=2.1.2
PyJWT~=2.4.0